Remove partially implemented support for empty regexes
With the full implementation it would have been possible to write these
regexes "", "()", "|", "(|)", "()|", "|()", "()|()" to match for example
the string "a" but not the string "ab". So it's like an empty move in
NFA. The regex engine PCRE supports such regexes.

The implementation would have made the regex parser more complicated
without providing any real added value in the regex power. For example,
the already supported syntax provides the same power: "a?" and "(a)?".
Moreover, this syntax is much more readable and less confusing.