Fix one-or-more NFA construction
The construction had no empty-move state as initial state.

With such a regex "ab+", the concatenation operation of NFA("a") and
NFA("b+") replaced an important state in NFA("b+"). Inside NFA("b+") a
state points to another state inside this NFA, but because of the
replacement the memory address of the state pointed to is no longer
correct.

The construction now has an empty-move state as initial state, which
fixes this problem.