Change the NFA/DFA to work byte-wise instead of character-wise
This change drastically reduces the required number of transitions that
one state must have when a wide range of characters must be matched.
It also removes PureBasic ASCII support.

Previously, a state had to be created with 65535 transitions to match
65535 different characters. DFA creation became extremely slow very
quickly when processing many NFA states at the same time that had a lot
of transitions (took 2-digit minutes quickly).

Now two states are needed to process one character, but this means that
each state only needs a maximum of 256 transitions. So the DFA creation
algorithm does not have to deal with too many transitions at the same
time and remains very fast.