Change incomplete DFA to a complete one
Previously, a hash map was used for the DFA state transitions (key was
the symbol; value was the next state) because each DFA state had only
the transitions that were really necessary and did not have transitions
for every symbol of the alphabet.

Now a complete DFA is created, meaning each state has transitions for
each symbol of the alphabet. In addition, the DFA now has a real dead
state. So, the DFA now meets the strict definition of a complete DFA.

This change allowed the hash map that stored the DFA state transitions
to be replaced with an array, which greatly increases the speed of the
DFA since there is now no need to compute hashes.