Add support to use an existing DFA stored in external memory
The DFA states are now accessed via a static array pointer, which now
allows an existing DFA stored in external memory to be used directly by
the RegEx engine. For example, a DFA can now be stored in a
`DataSection` block and the RegEx engine uses the DFA directly from
there.

- New function: `FreeDfa()`

- New function: `UseDfaFromMemory()`

- Added example code which uses a DFA stored in a `DataSection` block