Add support for custom character classes `[`...`]`
`[x]`, while `x` can be several mixes of the following:

- Symbol:
  - Single character
  - Escape sequence
- Range:
  - `a-c` for the characters `a`, `b` and `c`

To negate the user-defined character class, the `^` character must
follow immediately after the opening square bracket. If the `^` occurs
in other positions, it is a normal character, without the special
meaning.

The metacharacter `.` (dot) is a dot character inside the square
brackets, i.e. without the special meaning.

Nesting of character classes is not allowed.

For ranges, the range end symbol must have a larger character code than
the range start symbol, i.e. `[a-z]` instead of `[z-a]`.