Disable the PB debugger within the module
If predefined character classes are used in the RegEx, which are usually
very extensive, a very large NFA with very many states is quickly
required. The PB debugger slows down the processing within the module
extremely, so it is better to disable it within the module, so that the
module is also usable in PB debug mode.