Add feature to match multiple RegExes simultaneously
A RegEx ID number can now be specified for each RegEx, and in the event
of a match, the RegEx ID number can be used to determine which RegEx has
matched.

Useful for building a lexer where different RegEx ID numbers can be used
for the different token types.

- Function `Create(regExString$)` split into the functions:
  - `Init()`
  - `AddNfa(*regExEngine, regExString$, regExId = 0)`
- Function `Match()` has a new optional parameter:
  - `*regExId`

The `*regExId` parameter of the `Match()` function takes a pointer to an
integer variable into which the RegEx ID number of the matching RegEx is
written.

If there are multiple RegExes that match the same string and have been
assigned different RegEx ID numbers, the RegEx ID number of the last
matched RegEx is taken, i.e. the last matched RegEx added with the
`AddNfa()` function.