Add support for case-insensitive mode
New Syntax:

- `(?i)` means that everything after that is case-insensitive.
- `(?-i)` means that everything after that is case-sensitive.

New Parameter:

- `AddNfa(..., regExModes = 0)`

`AddNfa(..., "a", #RegExMode_NoCase)`
is the same as
`AddNfa(..., "(?i)a")`

Groups inherit the active modes of the context outside. Mode changes
within a group has no effect on the context outside this group.

The implementation uses Unicode's Simple Case Folding variant (single
character code point to single character code point), but in reverse:
Instead of mapping all character variations to a single character
(folding), a single character is mapped to all character variations
(unfolding). This is necessary because the DFA must know all valid
characters.

The source file from which the translation table is created:
https://www.unicode.org/Public/UCD/latest/ucd/CaseFolding.txt