Add reduced DFA matcher module
If only the precompiled DFAs are needed in the software for matching
and no new NFAs/DFAs are to be created at runtime, then the reduced
module can be used. This way the software is not unnecessarily
bloated with the large Unicode tables and the other code of the main
module.