Add ASCII mode
This feature implements ASCII mode. When activated, the predefined
character classes will only match the corresponding ASCII characters.
For example, `(?a)\w` will then match only `[a-zA-Z0-9_]`. The character
encoding remains UCS-2 in this mode, i.e. `(?a)\W` matches all UCS-2
characters, but not `[a-zA-Z0-9_]`.

New Syntax

- `(?a)` — activates ASCII mode.
- `(?-a)` — deactivates ASCII mode (default).

New Parameter

- `AddNfa(..., "\w", #RegExMode_Ascii)` is the same as
`AddNfa(..., "(?a)\w")`

This mode is also useful in combination with `#RegExMode_NoCase` when
you want to lex keywords in a code, case-insensitive, but no
case-folding should be applied. Example:

- `(?i)set` corresponds to `[Ss\u017F][Ee][Tt]`
- `(?ia)set` corresponds to `[Ss][Ee][Tt]`