Use AVL trees to speed up NFA matching and DFA creation
In the procedure `AddState()`, it is necessary to check whether a state
is already in a set of states in order to avoid adding the state
multiple times.

The set was stored as a list, which means that the list has to be
completely traversed each time to determine whether an entry already
exists.

With an AVL tree, it is not necessary to go through all the entries to
determine whether an entry already exists or not, so using an AVL tree
speeds up NFA matching and DFA creation.

Closes #40