Add single-byte mode
This feature allows to create a RegEx engine with a single-byte mode

When a RegEx engine is created with this mode, it will no longer
support all characters in UCS-2, but only those in ISO-8859-1
(unassigned areas are allowed, so `\x01` - `\xFF`.). This means that it
will only support UCS-2 characters where the second byte is zero. If
the RegEx contains characters outside ISO-8859-1, an error will occur.

This mode also activates the ASCII mode, which can't be deactivated
with `(?-a)` in the RegEx.

Since the second byte must always be zero, this check is then done in
the `Match` function. This way the NFA/DFA can save many states, making
it much smaller.

**New Parameter**
- `Init(RegEx::#RegExEngine_SingleByte)`

Closes #43