Revert "Use AVL trees to speed up NFA matching and DFA creation"
It was assumed that `AddElement()` always creates a list element whose
memory address is always higher than the previous list element, but
this is not the case, which is why the order of the `AddNfa()` calls
was not always observed during NFA/DFA processing. As a result, the
rule that, in the case of multiple `AddNfa()` calls whose RegEx can
match the same string, the `RegExId` of the last `AddNfa()` call is
taken was not always followed.